requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.1.3
django-bootstrap4==25.1
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
gunicorn==23.0.0
Pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.2
tzdata==2024.2
whitenoise==6.7.0